clean up shim build script
