--- packages/feedback/scripts/global-test.d.ts.tmpl
@@ -1,6 +1,0 @@
-declare module 'preact' {
-  export type VNode = any;
-  export type ComponentChildren = any;
-  export type ComponentType<T> = any;
-  export const h: any;
-}
